+++ macos/PaperMonitorApp/Sources/PaperMonitorApp/main.swift
@@ -4,5 +4,5 @@ import PaperMonitorCore
 let app = NSApplication.shared
 let delegate = AppDelegate(launchOptions: AppLaunchOptions())
 app.delegate = delegate
-app.setActivationPolicy(.regular)
+app.setActivationPolicy(.accessory)
 app.run()